+++ Runtime/Decal/NailEditor.cs
@@ -24,11 +24,11 @@ namespace net.rs64.TexTransTool.Decal
 
         internal override bool IsPossibleApply => TargetAvatar != null && TargetRenderers.Any(i => i != null);
 
-        internal override Dictionary<Material, RenderTexture> CompileDecal(ITextureManager textureManager, IIslandCache islandCacheManager, Dictionary<Material, RenderTexture> decalCompiledRenderTextures = null)
+        internal override Dictionary<Material, RenderTexture> CompileDecal(ITextureManager textureManager, Dictionary<Material, RenderTexture> decalCompiledRenderTextures = null)
         {
             decalCompiledRenderTextures ??= new();
 
-            foreach (var nailTexSpaceFilter in GetNailTexSpaceFilters(islandCacheManager))
+            foreach (var nailTexSpaceFilter in GetNailTexSpaceFilters())
             {
                 foreach (var renderer in TargetRenderers)
                 {
@@ -49,7 +49,7 @@ namespace net.rs64.TexTransTool.Decal
             return decalCompiledRenderTextures;
         }
 
-        List<(Texture2D, ParallelProjectionSpace, ParallelProjectionFilter<Vector2>)> GetNailTexSpaceFilters(IIslandCache islandCacheManager)
+        List<(Texture2D, ParallelProjectionSpace, ParallelProjectionFilter<Vector2>)> GetNailTexSpaceFilters()
         {
             var spaceList = new List<(Texture2D, ParallelProjectionSpace, ParallelProjectionFilter<Vector2>)>();
 
@@ -71,7 +71,7 @@ namespace net.rs64.TexTransTool.Decal
                     var islandSelector = new IslandSelector(new Ray(matrix.MultiplyPoint(Vector3.zero), matrix.MultiplyVector(Vector3.forward)), matrix.lossyScale.z * 1);
 
                     var SpaceConverter = new ParallelProjectionSpace(matrix.inverse);
-                    var Filter = new IslandCullingPPFilter<Vector2>(GetFilter(), new List<IslandSelector>(1) { islandSelector }, islandCacheManager);
+                    var Filter = new IslandCullingPPFilter<Vector2>(GetFilter(), new List<IslandSelector>(1) { islandSelector });
 
                     spaceList.Add((nailDecalDescription.DecalTexture, SpaceConverter, Filter));
                 }