+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Optimizing Phase が追加されました (#410)
 - GameObject から TexTransTool のほとんどのコンポーネントが追加できるようになりました (#411)
+- UVtoIsland の高速化 (#412)
 
 ### Changed
 
@@ -21,6 +22,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 ### Deprecated
+
+## [0.6.3]
+
+### Fixed
+
+- TexTransToolのコンポーネントを削除する処理で、GameObjectも削除されていた問題を修正 (#414)
 
 ## [0.6.2]
 
@@ -448,7 +455,8 @@ Cherry-Pick
 
 - AvatarTag 系のコンポーネントを削除しました `#11`
 
-[unreleased]: https://github.com/ReinaS-64892/TexTransTool/compare/v0.6.2...master
+[unreleased]: https://github.com/ReinaS-64892/TexTransTool/compare/v0.6.3...master
+[0.6.3]: https://github.com/ReinaS-64892/TexTransTool/compare/v0.6.2...v0.6.3
 [0.6.2]: https://github.com/ReinaS-64892/TexTransTool/compare/v0.6.1...v0.6.2
 [0.6.1]: https://github.com/ReinaS-64892/TexTransTool/compare/v0.6.0...v0.6.1
 [0.6.0]: https://github.com/ReinaS-64892/TexTransTool/compare/v0.5.7...v0.6.0